Add files via upload
The goal of this code is to replicate table number 5 of Perron's study titled "The Great Crash, The Oil Price Shock, and The Unit Root Hypothesis" (1989). This paper sought to prove that when taking into account structural breaks in macroeconomic series it is possible to reject the null hypothesis for unit root. However, some improvements should be made to accurately choose the lag length and the lambda value for each series. Also, this code only includes the models A and B. Any recommendations are welcome.